Kill StoRM webdav as same user that created it.
Motivation:

teapot will shutdown the StoRM WebDAV service as user root.  This is
unnecessary, as the targeted user is able to stop the StoRM WebDAV
service.

In general, running operations as root should be avoided.

Also, not requiring root makes it simpler to run teapot as a single user
(in a development environment).

Modification:

Update sudo command to target the targeted user.

Result:

teapot no longer makes use of root access to shutdown StoRM webdav.